+++ modules/oidc_token/get_audience.sh
@@ -1,10 +1,11 @@
-# Copyright 2019-2022 NephoSolutions srl, Sebastian Trebitz
+#!/bin/bash
+# Copyright 2022 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#     https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,20 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-terraform {
-  required_providers {
-    google = {
-      source  = "hashicorp/google"
-      version = ">= 3.52"
-    }
-    tfe = {
-      source  = "hashicorp/tfe"
-      version = ">= 0.36"
-    }
-    time = {
-      source  = "hashicorp/time"
-      version = ">= 0"
-    }
-  }
-  required_version = ">= 0.13"
+# Exit if any of the intermediate steps fail
+set -e
+
+cat <<EOF
+{
+  "audience": "$TFC_WORKLOAD_IDENTITY_AUDIENCE"
 }
+EOF